revertted changes due to recent mistakes
