+++ lib/players.rb
